check for collectors without network addresses
When using this cookbook with a ganglia server_role, the first node
launched will find itself as a collector after registering with
chef-server, but before returning network attributes. This means that
gmond_collectors will contain [nil] from the map, and .empty? will
return true. Using .any? instead checks for this failure case, falling
back to 127.0.0.1 until the next chef-client run, where it will be
updated with the local address.