prevent nil gmond_collectors addresses
When using this cookbook with a ganglia server_role, each node
launched will find itself as a collector after registering with
chef-server, but before returning network attributes. This means that
gmond_collectors will contain [nil] from the map, which confuses the
gmond.conf template with an empty value, preventing the service from
starting. .compact removes the nil values, and on the second chef-client
run, all correct addresses will be included.